Move RegisterEntity to Registers
